reward ko some management

+++ springJwt/src/main/java/com/authh/springJwt/Reward/Model/RewardModel.java
@@ -22,7 +22,7 @@ public class RewardModel {
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     private Long id;
  
-    private int rewardPoints;
+    private double rewardPoints;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
